fix: switch circle bug
